feat: Switch case feito para interface interativa

+++ soulpass_project/src/br/com/soulpass/models/ConversorPontos.java
@@ -1,7 +1,7 @@
 package br.com.soulpass.models;
 
 public class ConversorPontos {
-    public static double converterParaCredito(int pontos) {
+    public static double converterParaCredito(double pontos) {
         return pontos / 110.0;
     }
 }
